net: Move bridge NNCP fixture to shared l2_bridge conftest
Move bridge_nncp from vmi_interfaces_stability to the shared
l2_bridge conftest so it can be reused across l2_bridge test packages.
