+++ tests/network/l2_bridge/vmi_interfaces_stability/conftest.py
@@ -11,7 +11,6 @@ from tests.network.l2_bridge.vmi_interfaces_stability.lib_helpers import (
     secondary_network_vm,
     wait_for_stable_ifaces,
 )
-from utilities.constants import LINUX_BRIDGE, WORKER_NODE_LABEL_KEY
 
 
 @pytest.fixture(scope="class")
@@ -57,33 +56,6 @@ def bridge_nad(
         client=admin_client,
     ) as nad:
         yield nad
-
-
-@pytest.fixture(scope="class")
-def bridge_nncp(
-    nmstate_dependent_placeholder: None,
-    admin_client: DynamicClient,
-    hosts_common_available_ports: list[str],
-) -> Generator[libnncp.NodeNetworkConfigurationPolicy]:
-    with libnncp.NodeNetworkConfigurationPolicy(
-        client=admin_client,
-        name="iface-stability-bridge",
-        desired_state=libnncp.DesiredState(
-            interfaces=[
-                libnncp.Interface(
-                    name="br1-test",
-                    type=LINUX_BRIDGE,
-                    state=libnncp.Resource.Interface.State.UP,
-                    bridge=libnncp.Bridge(
-                        port=[libnncp.Port(name=hosts_common_available_ports[-1])],
-                    ),
-                )
-            ]
-        ),
-        node_selector={WORKER_NODE_LABEL_KEY: ""},
-    ) as nncp_br:
-        nncp_br.wait_for_status_success()
-        yield nncp_br
 
 
 @pytest.fixture(scope="class")
